Aedans gamesim

+++ gamesim/src/main/java/org/usfirst/irs1318/gamesim/units/Time.java
@@ -6,18 +6,15 @@ import java.util.Objects;
  * Class that represents an immutable instant in time.
  *
  * One unit of time is equivalent to one second in a game.
- *
- * Times should not be compared directly. Instead, they
- * should be checked to see if they're in a {@link Duration}.
  */
 public final class Time {
-    private final float x;
+    private final int x;
 
-    public Time(float x) {
+    public Time(int x) {
         this.x = x;
     }
 
-    public float getX() {
+    public int getX() {
         return x;
     }
